Vahak_01/requirements_02.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.3
django-environ==0.9.0
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mysqlclient==2.1.1
phonenumbers==8.13.2
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
